fix compile errors

+++ server/src/models/package_mysql.rs
@@ -20,7 +20,7 @@ pub struct DbPackage {
 
 #[derive(Serialize, Deserialize, Queryable, Debug, Clone)]
 pub struct Package {
-    pub id: i32,
+    pub id: u64,
     pub author: String,
     pub name: String,
     pub insert_date: Option<NaiveDateTime>,
